submit 1218 change return structure

+++ BackEnd/Mfapi/src/main/java/com/football/mfapi/dto/AnnouncementDTO.java
@@ -14,5 +14,5 @@ public class AnnouncementDTO {
     private String detail;
     private String title;
     private LocalDateTime time;
-    private Integer receiver;
+    private Long receiver;
 }
